requirements.txt
absl-py==0.14.0
aiohttp==3.7.4.post0
aircv==1.4.6
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
cnocr==2.0.1
colorama==0.4.4
fsspec==2021.9.0
future==0.18.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.41.0
idna==3.2
importlib-metadata==4.8.1
Markdown==3.3.4
MouseInfo==0.1.3
multidict==5.1.0
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.2
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pyDeprecate==0.3.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
pytorch-lightning==1.4.8
pytweening==1.0.4
pywin32==301
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.9.1
torchmetrics==0.5.1
torchvision==0.10.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.1

pandas~=1.3.4